# servants-of-india/docs/openapi.swagger.yaml
...
openapi: 3.0.3
info:
  title: Servants of Bharat API — Complete (All APIs)
  description: Swagger / OpenAPI 3.0 documentation for ALL APIs created for the Servants of Bharat platform.
    Every operation includes its mapped user story (x-user-story) and error-handling details (x-error-handling).
  version: 1.0.0
servers:
- url: /api
  description: Local API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Forbidden: insufficient role'
...
    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        volunteer_id:
          type: string
          format: uuid
        volunteer_name:
          type: string
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        event_id:
          type: string
          format: uuid
          nullable: true
        event_title:
          type: string
          nullable: true
        image_url:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - pending
          - approved
          - rejected
        submitted_at:
          type: string
          format: date-time
...
paths:
...
  /submissions:
    get:
      tags:
      - Submissions
      summary: Review queue (Event Manager sees own events only)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        schema:
          type: string
      x-user-story: As an event manager, I want to see the review queue of proofs for my events so that
        I can verify volunteer work.
      x-error-handling:
      - 401 - not authenticated
      - 403 - caller is not an Event Manager / Super Admin
      description: 'Review queue (Event Manager sees own events only)


        **User story:** As an event manager, I want to see the review queue of proofs for my events so
        that I can verify volunteer work.


        **Error handling:**

        - 401 - not authenticated

        - 403 - caller is not an Event Manager / Super Admin'
    post:
      tags:
      - Submissions
      summary: Create proof submission (Volunteer, multipart)
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - category_id
              - description
              - image
              properties:
                category_id:
                  type: string
                description:
                  type: string
                event_id:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Submission'
        '409':
          description: Duplicate/completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      x-user-story: As a volunteer, I want to submit proof of service for a completed event so that my
        work can be verified.
      x-error-handling:
      - 422 - missing category_id/description/image
      - 409 - already approved or a pending submission exists for this category
      - 413 - image larger than 5 MB
      - 401 - not authenticated
      description: 'Create proof submission (Volunteer, multipart)


        **User story:** As a volunteer, I want to submit proof of service for a completed event so that
        my work can be verified.


        **Error handling:**

        - 422 - missing category_id/description/image

        - 409 - already approved or a pending submission exists for this category

        - 413 - image larger than 5 MB

        - 401 - not authenticated'
  /submissions/me:
    get:
      tags:
      - Submissions
      summary: My submissions (Volunteer)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
      security:
      - bearerAuth: []
      x-user-story: As a volunteer, I want to see my submissions and their status so that I can track
        my progress.
      x-error-handling:
      - 401 - not authenticated
      - 403 - caller is not a Volunteer
      description: 'My submissions (Volunteer)


        **User story:** As a volunteer, I want to see my submissions and their status so that I can track
        my progress.


        **Error handling:**

        - 401 - not authenticated

        - 403 - caller is not a Volunteer'
  /submissions/{id}:
    get:
      tags:
      - Submissions
      summary: Get a submission
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Submission'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      x-user-story: As a reviewer or owner, I want to view a submission's details so that I can assess
        it.
      x-error-handling:
      - 404 - submission not found
      - 403 - not the owner or a reviewer
      - 401 - not authenticated
      description: 'Get a submission


        **User story:** As a reviewer or owner, I want to view a submission''s details so that I can assess
        it.


        **Error handling:**

        - 404 - submission not found

        - 403 - not the owner or a reviewer

        - 401 - not authenticated'
...